Add way to run protractor tests w/o starting services
Also add pouch mapreduce because
