+++ scripts/e2e-servers
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+const moduleManager = require('../tests/protractor/module-manager');
+
+Promise.resolve()
+  .then(() => moduleManager.startNodeModule('api'))
+  .then(() => moduleManager.startNodeModule('sentinel'));
