+++ tests/protractor/tests-only.conf.js
@@ -1,0 +1,2 @@
+module.exports = require('./conf');
+module.exports.manageModules = false;
